Perfrom BLAST Searches Against a Set of Genomes

This function takes a fasta file containing query sequences as input and performs BLAST searches of these query sequences against a set of reference genomes to retrieve corresponding hits in diverse genomes.

blast_nucleotide_to_genomes(
  query,
  subject_genomes,
  blast_type = "blastn",
  blast_output_path = "blast_output",
  min_alig_length = 30,
  evalue = 1e-05,
  max.target.seqs = 5000,
  update = FALSE,
  ...
)

Arguments

query

path to input file in fasta format.

subject_genomes

a vector containing file paths to the reference genomes that shall be queried (e.g. file paths returned by meta.retrieval).

blast_type

specification of the BLAST type shall be used to perform BLAST searches between query and reference. Available options are:

blast_type = "blastn" for nucleotide to nucleotide BLAST searches (see blast_nucleotide_to_nucleotide for details)

blast_type = "tblastn" for protein to nucleotide BLAST searches (see blast_protein_to_nucleotide for details)

blast_output_path

a path to a folder that will be created to store BLAST output tables for each individual query-genome search.

min_alig_length

minimum alignment length that shall be retained in the result dataset. All hit alignments with smaller hit alignment length will be removed automatically.

evalue

minimum expectation value (E) threshold for retaining hits (default: evalue = 0.00001).

max.target.seqs

maximum number of aligned sequences that shall be retained. Please be aware that max.target.seqs selects best hits based on the database entry and not by the best e-value. See details here: https://academic.oup.com/bioinformatics/advance-article/doi/10.1093/bioinformatics/bty833/5106166 .

update

a logical value indicating whether or not pre-computed BLAST tables should be removed and re-computed (update = TRUE) or imported from existing file (update = FALSE) (Default).

...

additional arguments passed to blast_nucleotide_to_nucleotide or blast_protein_to_nucleotide.

Details

The blast_nucleotide_to_genomes function enables users to BLAST specific query sequences against a set of reference genomes and retrieve the corresponding BLAST output.

Author

Hajk-Georg Drost