Perform Reverse PSI-BLAST searches (rpsblast)

Run rpsblast (Reverse PSI-BLAST) searches a query sequence against a database of profiles, or score matrices, producing BLAST-like output.

blast_rpsblast(
  query,
  db,
  db.alias,
  prep.db = FALSE,
  output.path = NULL,
  db.import = FALSE,
  postgres.user = NULL,
  evalue = 0.001,
  out.format = "csv",
  cores = 1,
  blast.path = NULL
)

Arguments

query

path to input file in fasta format.

db

path to rpsblast-able database or smp files if prep.db = NULL.

db.alias

alias for database files

prep.db

logical, default FALSE; if TRUE - prep db before running rpsblast.

output.path

path to folder at which BLAST output table shall be stored.

db.import

shall the BLAST output be stored in a PostgresSQL database and shall a connection be established to this database? Default is db.import = FALSE. In case users wish to to only generate a BLAST output file without importing it to the current R session they can specify db.import = NULL.

postgres.user

when db.import = TRUE and out.format = "postgres" is selected, the BLAST output is imported and stored in a PostgresSQL database. In that case, users need to have PostgresSQL installed and initialized on their system. Please consult the Installation Vignette for details.

evalue

Expectation value (E) threshold for saving hits (default: evalue = 0.001).

out.format

a character string specifying the format of the file in which the rpsblast results shall be stored. Available options are:

out.format = "pair" : Pairwise

out.format = "qa.ident" : Query-anchored showing identities

out.format = "qa.nonident" : Query-anchored no identities

out.format = "fq.ident" : Flat query-anchored showing identities

out.format = "fq.nonident" : Flat query-anchored no identities

out.format = "xml" : XML

out.format = "tab" : Tabular separated file

out.format = "tab.comment" : Tabular separated file with comment lines

out.format = "ASN.1.text" : Seqalign (Text ASN.1)

out.format = "ASN.1.binary" : Seqalign (Binary ASN.1)

out.format = "csv" : Comma-separated values

out.format = "ASN.1" : BLAST archive (ASN.1)

out.format = "json.seq.aln" : Seqalign (JSON)

out.format = "json.blast.multi" : Multiple-file BLAST JSON

out.format = "xml2.blast.multi" : Multiple-file BLAST XML2

cores

number of cores for parallel rpsblast searches.

blast.path

path to rpsblast executables.

See also

Author

Hajk-Georg Drost, Anna Gogleva

Examples

if (FALSE) {
db_path <- '/home/Cdd'
db_alias <- 'myCdd'
rps_test <- blast_rpsblast(query = system.file('seqs/sbj_aa.fa', package = 'metablastr'), 
                           db = db_path, 
                           db.alias = db_alias, 
                           prep.db = FALSE)
}