Extract sequences of BLAST hits in respective genomes and store it as fasta file(s)

In some cases, users may wish to extract the sequences of the respective blast hit stored in a blast_tbl. This function enables to quickly extract such sequences and store them in one general or multiple genome specific fasta file(s).

extract_hit_seqs_from_genomes(
  blast_tbl,
  subject_genomes,
  file_name = NULL,
  separated_by_genome = FALSE,
  update = TRUE,
  path = NULL
)

Arguments

blast_tbl

a BLAST table generated with blast_genomes.

subject_genomes

a vector containing file paths to the reference genomes that shall be queried (e.g. file paths returned by meta.retrieval).

file_name

name of the fasta file that stores the BLAST hit sequences. This name will only be used when separated_by_genome = FALSE.

separated_by_genome

a logical value indicating whether or not hit sequences from different genomes should be stored in the same output fasta file separated_by_genome = FALSE (default) or in separate fasta files separated_by_genome = TRUE.

update

shall an existing file_name file be overwritten (update = TRUE; Default) or shall blast hit sequences be appended to the existing file (update = FALSE)?

path

a folder path in which corresponding fasta output files shall be stored.

Author

Hajk-Georg Drost