Plot pairwise dN, dS, or dN/dS distributions of orthologous genes

This function enables users to plot pairwise dN, dS, or dN/dS distributions of orthologous genes.

gg_species_dnds_blast_tbl(
  blast_tbl,
  type,
  order,
  xlab = "",
  ylab = "",
  title = ""
)

Arguments

blast_tbl

a blast_tbl.

type

type of sustitution rate quantification that shall be visualized on the y-axis. Options are:

type = "dN"

type = "dS"

type = "dN/dS"

type = "dN+dS"

order

a character vector specifying the order of the species using their scientific names (which have to match the species names stored in the subject_species column of the input blast_tbl).

xlab

label of x-axis.

ylab

label of y-axis.

title

title label.

Author

Hajk-Georg Drost