Count the number of motifs in a set of non-random versus random sequences

Compare the number of motifs in a set of non-random versus random sequences. The resulting values can then be used to test the enrichment of certain motifs in real sequences compared to random sequences.

motif_compare(
  real_seqs,
  random_seqs,
  motifs,
  max.mismatch = 0,
  min.mismatch = 0,
  ...
)

Arguments

real_seqs

a file path to the fasta file storing the non-random set of sequences.

random_seqs

a file path to the fasta file storing the random set of sequences, e.g. generated with extract_random_seqs_from_genome.

motifs

a character vector storing a set of motifs that shall be counted within respective sequences.

max.mismatch

the maximum number of mismatching letters allowed (see matchPattern for details).

min.mismatch

the minimum number of mismatching letters allowed (see vcountPattern for details).

...

additional arguments passed to matchPattern.

See also

Author

Hajk-Georg Drost