Count the number of motifs in a set of sequences

Detect all matches of a given motif (typically a short string) in a (typically long sequence) reference sequence or set of reference sequences.

motif_count(file, motif, max.mismatch = 0, min.mismatch = 0, ...)

Arguments

file
a file path to the fasta file storing the sequences of interest.

motif
a character string specifying the sequence motif that shall be searched.

max.mismatch
the maximum number of mismatching letters allowed (see matchPattern for details).

min.mismatch
the minimum number of mismatching letters allowed (see vcountPattern for details).

...
additional arguments passed to matchPattern.

Author

Hajk-Georg Drost