Import BLAST output file

When performing BLAST searches with the blast_*() functions, the corresponding BLAST output file can be imported into the current R session using this function. All output formats given by BLAST are supported (see e.g. description for details).

read_blast(file, out.format, postgres.user = NULL)

Arguments

file
path to BLAST output file.

out.format
a character string specifying the output format of the BLAST output that shall be imported. Available options are:

out.format = "postgres" : store BLAST output as Postgres database and generate postgres database connection.

out.format = "xml" : XML

out.format = "tsv" : Tabular separated file

out.format = "csv" : Comma-separated values

out.format = "json.seq.aln" : Seqalign (JSON)

out.format = "json.blast.multi" : Multiple-file BLAST JSON

out.format = "xml2.blast.multi" : Multiple-file BLAST XML2

out.format = "json.blast.single" : Single-file BLAST JSON

out.format = "xml2.blast.single" : Single-file BLAST XML2

@param postgres.user specify username for RPostgreSQL connection. @author Hajk-Georg Drost @seealso blast_protein_to_protein, blast_nucleotide_to_protein, blast_nucleotide_to_nucleotide @importFrom RPostgreSQL @export